requirements.txt
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.7.0
joblib==1.4.2
prefect==3.4.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.12
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
streamlit==1.38.0
